dashboard: split dashboard web server code into its own module
